feat(dsv4): tnd-opus55 initial — precision fix + ACLGraph alignment + CSA replay gate
精度修复（最高优先级）：
- pto_attn.py: cmp_norm_w / inner_norm_w dtype FP32 → BF16
  compressor 和 inner compressor 的 RMS norm weight 在加载后保持 BF16，
  tnd 分支错误验证为 FP32，导致 kernel RMS 路径静默精度损失。
  nalinaly 分支同位置保持 BF16，与此修复一致。

ACLGraph 捕获档位对齐：
- platform.py: _setup_compile_backend 新增档位 round-up 到 6 的倍数
  DSpark 每请求 6 token（5 spec + 1 draft），BSH layout kernel 要求 T = B×6。
  非 6 倍数档位导致 kernel token axis 不整除，同时是精度对比的前置条件。
  触发条件：speculative_config 存在 + cudagraph 开启 + 非 eager。
  可通过 additional_config.ascend_compilation_config.align_decode_capture_sizes=False 关闭。

新增 service_config.py：
- pto_kernels/dspark/service_config.py: QUERY_TOKENS=6, can_replay_csa_graph()
  图重放判断：档位在 CSA 范围内时要求 uniform decode 且每请求恰好 6 token；
  范围外放行，不影响非 DSpark 场景。
  本机 CPU 合约：7/7 PASS。

model_runner_v1.py 两处新增：
- dispatch_cudagraph 之后：CSA 图重放闸门
  非 uniform batch 降回 CUDAGraphMode.NONE，防止补位档位被非法 replay。
- get_model() 之后：process_weights_after_loading 主动调用
  vllm-ascend main 无 model-level 自动调用入口；CSA kernel 的 Hadamard 矩阵
  转换和 INT8 scale 绑定必须在图捕获 warmup 之前完成，防止捕获流中出现
  device→host 操作。

测试清单：tests/pypto_test/TEST_CHECKLIST_TND_OPUS55.md
设计文档：reports/dsv4-tnd-opus55-20260924/design.md（独立 reports 目录）

+++ vllm_ascend/attention/pto_attn.py
@@ -167,7 +167,7 @@ def prepare_weights(impl, hadamard):
         "cmp_wkv": _dense(compressor.wkv, (kcsa.MAIN_OUT_DIM, D)),
         "cmp_wgate": _dense(compressor.wgate, (kcsa.MAIN_OUT_DIM, D)),
         "cmp_ape": _native_tensor("cmp_ape", compressor.ape.detach(), torch.float32),
-        "cmp_norm_w": _native_tensor("cmp_norm_w", compressor.norm.weight.detach(), torch.float32),
+        "cmp_norm_w": _native_tensor("cmp_norm_w", compressor.norm.weight.detach(), torch.bfloat16),
         "idx_wq_b": idx_wq_b,
         "idx_wq_b_scale": idx_wq_b_scale,
         "weights_proj": _dense(indexer.weights_proj, (D, IH)),
@@ -177,7 +177,7 @@ def prepare_weights(impl, hadamard):
         "inner_wkv": _dense(indexer.compressor.wkv, (kcsa.INNER_OUT_DIM, D)),
         "inner_wgate": _dense(indexer.compressor.wgate, (kcsa.INNER_OUT_DIM, D)),
         "inner_ape": _native_tensor("inner_ape", indexer.compressor.ape.detach(), torch.float32),
-        "inner_norm_w": _native_tensor("inner_norm_w", indexer.compressor.norm.weight.detach(), torch.float32),
+        "inner_norm_w": _native_tensor("inner_norm_w", indexer.compressor.norm.weight.detach(), torch.bfloat16),
         "attn_sink": _native_tensor("attn_sink", impl.attn_sink.detach(), torch.float32),
         # vLLM keeps [G, O_GROUP_IN, O_LORA]; the kernel wants the transpose.
         "wo_a": _native_tensor("wo_a", impl.wo_a.weight.detach(), torch.bfloat16).transpose(1, 2).contiguous(),
